Merged revision(s) 518 from branches/SFML2:
Bugfixes:
- Slightly improved error handling (not fully merged from branches/SFML2)
- Fixed Credits GUI broken in r492
- Updated some copyright dates to 2014

Refactorizations:
- Fixed some warnings
- Avoid copying TScore objects
- Small simplifications
........

Fixed mistake in previous merge commit.

+++ trunk/src/event_select.cpp
@@ -103,7 +103,7 @@ void CEventSelect::Enter() {
 
 	int siz = FT.AutoSizeN(5);
 
-	double len = FT.GetTextWidth(Trans.Text(9));
+	float len = FT.GetTextWidth(Trans.Text(9));
 	textbuttons[0] = AddTextButton(Trans.Text(9), area.right-len-50, AutoYPosN(70), siz);
 	textbuttons[1] = AddTextButton(Trans.Text(8), area.left+50, AutoYPosN(70), siz);
 	SetFocus(textbuttons[1]);
